Remove hearing silent hearing dependency from RHS compats
