+++ addons/compat_rhs_afrf3/compat_rhs_afrf3_hearing/CfgWeapons.hpp
@@ -1,0 +1,28 @@
+class CfgWeapons {
+    class H_HelmetB;
+    class rhs_tsh4: H_HelmetB {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class rhs_zsh7a: H_HelmetB {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class rhs_gssh18: H_HelmetB {
+        HEARING_PROTECTION_EARMUFF;
+    };
+    class rhs_6m2: H_HelmetB {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class rhs_6b47;
+    class rhs_6b47_6m2: rhs_6b47 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhs_6b47_6m2_1: rhs_6b47 {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class rhs_6b47_bare;
+    class rhs_6b48: rhs_6b47_bare {
+        HEARING_PROTECTION_VICCREW;
+    };
+};
