+++ addons/compat_rhs_afrf3/config.cpp
@@ -1,5 +1,4 @@
 #include "script_component.hpp"
-#include "\z\ace\addons\hearing\script_macros_hearingProtection.hpp"
 
 class CfgPatches {
     class ADDON {
