+++ addons/compat_rhs_usf3/compat_rhs_usf3_hearing/CfgWeapons.hpp
@@ -1,0 +1,174 @@
+class CfgWeapons {
+    // Fast Helmets
+    class rhsusf_opscore_01;
+    class rhsusf_opscore_ut_pelt_nsw: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_aor1_pelt: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_aor1_pelt_nsw: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_bk_pelt: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_fg_pelt: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_fg_pelt_nsw: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_fg_pelt_cam: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_paint_pelt: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_paint_pelt_nsw: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_paint_pelt_nsw_cam: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_aor2_pelt: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_aor2_pelt_nsw: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_ut_pelt: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_ut_pelt_cam: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_ut_pelt_nsw_cam: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_mc_pelt: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_mc_pelt_nsw: rhsusf_opscore_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_cover;
+    class rhsusf_opscore_mc_cover_pelt_nsw: rhsusf_opscore_cover {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_mc_cover_pelt: rhsusf_opscore_cover {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_mc_cover_pelt_cam: rhsusf_opscore_cover {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_rg_cover_pelt: rhsusf_opscore_cover {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_coy_cover_pelt: rhsusf_opscore_cover {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_mar_01;
+    class rhsusf_opscore_mar_ut_pelt: rhsusf_opscore_mar_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_opscore_mar_fg_pelt: rhsusf_opscore_mar_01 {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    // ACH Helmets
+    class rhsusf_ach_helmet_ocp;
+    class rhsusf_ach_bare_des_headset: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_bare_des_headset_ess: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_bare_headset: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_bare_headset_ess: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_bare_semi_headset: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_bare_semi_headset_ess: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_bare_tan_headset: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_bare_tan_headset_ess: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_bare_wood_headset: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_bare_wood_headset_ess: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_helmet_headset_ocp: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_ach_helmet_headset_ess_ocp: rhsusf_ach_helmet_ocp {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    // ACVC Helmets
+    class rhsusf_cvc_helmet: rhsusf_opscore_01 {
+        HEARING_PROTECTION_VICCREW;
+    };
+
+    // MICH Helmets
+    class rhsusf_mich_bare;
+    class rhsusf_mich_bare_alt: rhsusf_mich_bare {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_mich_bare_norotos;
+    class rhsusf_mich_bare_norotos_alt: rhsusf_mich_bare_norotos {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_mich_bare_norotos_arc;
+    class rhsusf_mich_bare_norotos_arc_alt: rhsusf_mich_bare_norotos_arc {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_mich_bare_semi;
+    class rhsusf_mich_bare_alt_semi: rhsusf_mich_bare_semi {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_mich_bare_norotos_semi;
+    class rhsusf_mich_bare_norotos_alt_semi: rhsusf_mich_bare_norotos_semi {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_mich_bare_norotos_arc_semi: rhsusf_mich_bare_norotos_alt_semi {
+        HEARING_PROTECTION_OPEN;
+    };
+    class rhsusf_mich_bare_norotos_arc_alt_semi: rhsusf_mich_bare_norotos_arc_semi {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_mich_bare_tan;
+    class rhsusf_mich_bare_alt_tan: rhsusf_mich_bare_tan {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_mich_bare_norotos_tan;
+    class rhsusf_mich_bare_norotos_alt_tan: rhsusf_mich_bare_norotos_tan {
+        HEARING_PROTECTION_PELTOR;
+    };
+    class rhsusf_mich_bare_norotos_arc_tan;
+    class rhsusf_mich_bare_norotos_arc_alt_tan: rhsusf_mich_bare_norotos_arc_tan {
+        HEARING_PROTECTION_PELTOR;
+    };
+
+    class rhsusf_hgu56p: rhsusf_opscore_01 {
+        HEARING_PROTECTION_VICCREW;
+    };
+    class rhsusf_ihadss: rhsusf_opscore_01 {
+        HEARING_PROTECTION_VICCREW;
+    };
+
+    class H_HelmetB;
+    class RHS_jetpilot_usaf: H_HelmetB {
+        HEARING_PROTECTION_VICCREW;
+    };
+};
