+++ addons/compat_rhs_usf3/compat_rhs_usf3_hearing/script_component.hpp
@@ -1,0 +1,3 @@
+#define SUBCOMPONENT hearing
+#define SUBCOMPONENT_BEAUTIFIED Hearing
+#include "..\script_component.hpp"
